ajustando task de download part 2
