+++ pipelines/datasets/br_anatel_banda_larga_fixa/tasks.py
@@ -17,31 +17,10 @@ from pipelines.datasets.br_anatel_banda_larga_fixa.constants import (
 )
 from pipelines.datasets.br_anatel_banda_larga_fixa.utils import (
     check_and_create_column,
-    download,
+    descompactar_arquivo,
     to_partitions_microdados,
 )
-from pipelines.utils.utils import log, to_partitions
-
-@task(
-    max_retries=20,
-    retry_delay=timedelta(seconds=constants.TASK_RETRY_DELAY.value),
-)
-def descompactar_arquivo():
-    download(path = anatel_constants.INPUT_PATH.value)
-    # Obtenha o nome do arquivo ZIP baixado
-    zip_file_path = os.path.join(anatel_constants.INPUT_PATH.value, 'acessos_banda_larga_fixa.zip')
-    time.sleep(150)
-    print(os.listdir(anatel_constants.INPUT_PATH.value))
-    try:
-        with ZipFile(zip_file_path, 'r') as zip_ref:
-            zip_ref.extractall(anatel_constants.INPUT_PATH.value)
-
-    except Exception as e:
-            print(f"Erro ao baixar ou extrair o arquivo ZIP: {str(e)}")
-
-
-    os.remove(zip_file_path)
-    return print(os.listdir(anatel_constants.INPUT_PATH.value))
+from pipelines.utils.utils import log
 
 
 @task(
